# Prioritize short inference, unmerge pools for long inference