--- a/tests/test_block_exact_match.py
+++ b/tests/test_block_exact_match.py
@@ -4,6 +4,7 @@ import pytest
 import torch
 
 from petals import AutoDistributedConfig, RemoteSequential
+from petals.server.block_functions import MAX_SHORT_INFERENCE_TOKENS
 from petals.server.from_pretrained import load_pretrained_block
 from test_utils import *
 
@@ -16,13 +17,17 @@ def test_remote_block_exact_match(atol_forward=1e-4, atol_inference=1e-3):
     for block_index in random.sample(range(config.num_hidden_layers), 3):
         remote_block = remote_sequential[block_index]
 
-        inputs = torch.randn(1, 8, config.hidden_size)
+        inputs = torch.randn(1, MAX_SHORT_INFERENCE_TOKENS + 8, config.hidden_size)
         outputs_forward = remote_block(inputs)
 
         outputs_inference = []
         with torch.inference_mode():
             with remote_block.inference_session(max_length=inputs.shape[1]) as sess:
-                for i in range(inputs.shape[1]):
+                # Test long inference (unmerged inference pools)
+                outputs_inference.append(sess.step(inputs[:, : MAX_SHORT_INFERENCE_TOKENS + 1, :]))
+
+                # Test short inference (merged inference pools)
+                for i in range(MAX_SHORT_INFERENCE_TOKENS + 1, inputs.shape[1]):
                     outputs_inference.append(sess.step(inputs[:, i : i + 1, :]))
 
                 # test that max length is respected