Right now, long inference requests may occupy Runtime for a few seconds without giving it away to process short (most latency-sensitive requests). This PR fixes it by disallowing the merged pool for long requests and prioritizing the short ones.

**Tests:**

- 1 token: 130 ms => 130 ms
- 16 tokens: 400 ms => 500 ms
- 64 tokens: 450 ms => 550 ms
- 2048 tokens: ? => 5.2 s

This slowdown is more acceptable than occupying the runtime for the duration from 400 ms 5-20 s (for Llama 2, 8192 tokens) in case of prefix processing.